# Add bufreader example