Trying to capture just how effective a bufreader is for the BIP-324 protocol. I think this shows that it is usually going to help, probably never hurt.

How it generally looks on my machine:

```
> cargo run --release --example bufreader --features tokio
    Finished `release` profile [optimized] target(s) in 0.05s
     Running `target/release/examples/bufreader`
Bitcoin Traffic: 100000 total messages, 9940000 bytes
  Without BufReader: 126 ms
  With BufReader:    98 ms
  Improvement:       22.1%
Large Messages: 3000 total messages, 90112000 bytes
  Without BufReader: 234 ms
  With BufReader:    233 ms
  Improvement:       0.1%
Small Messages: 100000 total messages, 300000 bytes
  Without BufReader: 101 ms
  With BufReader:    72 ms
  Improvement:       28.2%
```